fix(db5): Move home, signin, signup pages to new folder (routes)
